fix: random fatal error failures during teardown
The issue is that this test was checking the reception of a
sample during start, but not the emission of the start event.
This caused the test's own teardown to not explicitly stop the
task, which left it to Syskit's own teardown procedure. During
which the "fake modbus" driver was not present.