fix: make sure the task does not enter an infinite loop if "ghost" bytes arrive
If at least one "ghost" byte arrives, that is one that is not part of the
master/slave conversation, the iodrivers_base::Task::updateHook calls
processIO, which does nothing in this task, and then we enter an infinite loop.

The original task was calling updateHook last to avoid this problem - by chance
or by design, but with no comments. This got "normalized" in decembr of 2024
to moving it at the top.

To avoid this dependency on call order, this commit insteads throws in
processIO.